+++ apps/native-component-list/src/screens/ExpoCoreModuleScreen.tsx
@@ -1,4 +1,4 @@
-import { reloadAsync } from 'expo';
+import { reloadAppAsync } from 'expo';
 import { ScrollView, StyleSheet } from 'react-native';
 
 import ListButton from '../components/ListButton';
@@ -7,9 +7,9 @@ export default function ExpoCoreModuleScreen() {
   return (
     <ScrollView style={styles.scrollView}>
       <ListButton
-        title="reloadAsync"
+        title="reloadAppAsync"
         onPress={() => {
-          reloadAsync('Reload from ExpoCoreModuleScreen');
+          reloadAppAsync('Reload from ExpoCoreModuleScreen');
         }}
       />
     </ScrollView>