+++ packages/expo-modules-core/src/ts-declarations/global.ts
@@ -46,7 +46,7 @@ export interface ExpoGlobal {
   /**
    * Reloads the app.
    */
-  reloadAsync(reason: string): Promise<void>;
+  reloadAppAsync(reason: string): Promise<void>;
 }
 
 type ViewConfig = {