+++ packages/expo-modules-core/src/web/CoreModule.ts
@@ -80,7 +80,7 @@ globalThis.expo = {
   getViewConfig: () => {
     throw new Error('Method not implemented.');
   },
-  reloadAsync: async () => {
+  reloadAppAsync: async () => {
     window.location.reload();
   },
 };